Create README.md
Afaq Ashfaq is a passionate AI researcher and developer with expertise in Machine Learning, Explainable AI (XAI), and Deep Learning. With an MSc in Artificial Intelligence from the University of Stirling, he specializes in projects at the intersection of technology and healthcare, including Natural Language Processing in radiology and GPU-based sensitivity estimation. Dedicated to advancing AI for practical and impactful applications, Afaq is always eager to collaborate on cutting-edge projects, contribute to open-source initiatives, and explore innovative solutions for real-world challenges.